scripts/objectron/requirements.txt
opencv-python
Pillow
tqdm
google==3.0.0
google-api-core==2.19.2
google-api-python-client==2.144.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
